Move the interpreter for the command-line interface into its module.
