+++ app/Main.hs
@@ -1,16 +1,10 @@
 module Main where
 
-import Data.Foldable (for_, toList, traverse_)
-import Data.Result
 import Data.Semigroup
 import Data.Version (showVersion)
 import Options.Applicative
-import Parser
 import qualified Paths_surface as Library (version)
-import qualified REPL
 import Surface.Command
-import Surface.Proof
-import Text.Pretty
 
 command :: Parser Command
 command
@@ -26,27 +20,7 @@ arguments = info
   <> header "surface - a dependently typed language with nothing much to say for itself")
 
 main :: IO ()
-main = do
-  command <- execParser arguments
-  case command of
-    Interactive -> REPL.runREPL REPL.repl
-    Run path -> do
-      result <- parseFromFile source path
-      printResult $ do
-        modules <- result
-        for_ modules (run . checkModule)
-    Debug path -> do
-      result <- parseFromFile source path
-      traverse_ (traverse_ prettyPrint) $ do
-        modules <- result
-        return $ do
-          m <- toList modules
-          runSteps initialState (checkModule m)
-
-printResult :: Pretty a => Result a -> IO ()
-printResult result = case result of
-  Result a -> prettyPrint a
-  Error es -> for_ es putStr
+main = execParser arguments >>= runCommand
 
 versionString :: String
 versionString = "Surface version " <> showVersion Library.version
